Small fixes to run.sh

+++ docker/zusam/run.sh
@@ -10,7 +10,7 @@ echo "* * * * * su-exec ${UID}:${GID} /zusam/api/bin/console zusam:cron >> /zusa
 
 DATABASE_URL="sqlite:///%kernel.project_dir%/../data/${DATABASE_NAME}"
 
-if [ ${INSTANCE_TYPE} == "demo" ]; then
+if [ "${INSTANCE_TYPE}" = "demo" ]; then
     ENV="dev"
 else
     ENV="prod"
@@ -25,14 +25,16 @@ sed -i -e "s|<SECRET>|$(openssl rand -base64 48)|g" \
 if ! [ -f /zusam/data/config.yml ]; then
     cp /zusam/config.yml /zusam/data/config.yml
 fi
-ln -s /zusam/data/files /zusam/public/files
+if ! [ -L /zusam/public/files ]; then
+    ln -s /zusam/data/files /zusam/public/files
+fi
 
-if [ ${INSTANCE_TYPE} == "demo" ] || [ ${INSTANCE_TYPE} == "new" ]; then
+if [ "${INSTANCE_TYPE}" = "demo" ] || [ "${INSTANCE_TYPE}" = "new" ]; then
     reset.sh
 fi
 
 # if this is a demo instance, reset it every day
-if [ ${INSTANCE_TYPE} == "demo" ]; then
+if [ "${INSTANCE_TYPE}" = "demo" ]; then
     echo "0 0 * * * su-exec ${UID}:${GID} /usr/local/bin/reset.sh >> /zusam/api/var/log/cron.log" | crontab -
 fi
